Propagate @Deprecated to class members in removed files
Previously, the deprecated status was treated differently in
`*removed.txt` and `*current.txt`. In the former deprecating the class
did not result in `@Deprecated` being added recursively to class
members and nested classes but in the latter it did. The reason for
that discrepancy was simply due to an inconsistency in the way it had
been implemented and not for some wider purpose. See change
https://r.android.com/3121577 for more details.

This change removes that behavior and follow-on changes will update the
affected `*removed.txt` signature files in Android. The change broke
two tests, one in `ApiAnalyzerTest` and one in `DeprecatedTest`. Rather
than spend time fixing both the former was deleted as the latter is a
more comprehensive test and the test class is focused on testing
deprecation behavior.

Bug: 373584764
Test: ./gradlew
Change-Id: I1191e4439ce699f8e99e1942b2af0b0617a2f046