Include property annotations even when the property has a backing field
Previously, annotations on properties with the `property:` target or no target were only being included on `PsiPropertyItem`s when the property did not have a backing field. This adds the annotations when the backing field exists as well -- the backing field might be private, and even when the field is in the signature file it seems inconsistent to only sometimes include property annotations.

This results in AndroidX API changes: aosp/3315190

Bug: 369846185
Test: Updated existing tests with property annotations
Change-Id: Ib3ab0d80d7a5ca6595cf78e3af240bd1c50bd9f6