Filter out no-args value class primary constructor
When a primary constructor in Kotlin has a single optional arg, a no-args constructor is generated for Java interop, even if JvmOverloads is not used. However, value class constructors cannot be called from Java, so there should not be a no-args constructor generated in that case. Currently K1 and K2 create the fake constructor, but going forward K1 will not, so this change stops generating it for consistency.

This results in AndroidX API changes: aosp/3314852

Bug: 369846185
Test: CommonValueClassTest
Change-Id: Ifaa91593197644e73b794d4f6f62c12d4c13d89f