Avoid ambiguity around phrase "current API surface"
Previously, `SignatureFile` had a `forCurrentApiSurface` property that
indicated whether the file contained definitions of APIs that belonged
to the API currently being generated by Metalava. e.g. If Metalava was
generating the system API then `forCurrentApiSurface` would be `false`
for the `..../current.txt` file containing the public API but `true`
for the `..../system-current.txt` file containing the system API delta.

Unfortunately, "current" has multiple meanings which makes the use
ambiguous, e.g.
* It could mean the `current.txt` file.
* It could mean the `prebuilts/sdk/<n>/public/api/<module>.txt` file.
* It could mean the whole latest API.
* It could mean the main API being generated by Metalava.

This change resolves that ambiguity by renaming the
`...CurrentApiSurface...` symbols to `...MainApiSurface...` and
updating the comments to match.

Bug: 374096206
Test: ./gradlew
Change-Id: I7f4c6ca63d7bf3da59c8f40195733cb917940073