Remove MutableCodebase, use DefaultCodebase instead
The `MutableCodebase` class was added as an intermediate step during
the work to make `PsiBasedCodebase` extend `DefaultCodebase`. It
allowed the work to make `PsiClassItem` extend `DefaultClassItem` to be
done separately. Now that `PsiBasedCodebase` extends `DefaultCodebase`
it is no longer needed so this change removes it and replaces the
reference with `DefaultCodebase`

Bug: 352480646
Test: ./gradlew
Change-Id: I8c510d8073b449937f5346d022666d22ed007195