Fix tests that do not configure API surfaces correctly
A follow-up change will make the parsing of signature files stricter
by requiring that if multiple signature files are provided that
contain base API plus deltas on top, e.g. public and system, that the
test is configured to support both the `main` and `base` `ApiSurface`s.

This change fixes up the tests that are affected by that change as
follows:
* The `ApiFileTest` is updated to explicitly create an `ApiSurfaces`
  that has a `base` `ApiSurface` rather than rely on the default
  `ApiSurfaces` which does not.

* The `MultipleCompatibilityFilesTest` is updated to specify the
  necessary `--show-annotation ...SystemApi` command line options so
  that the `ApiSurfaces` configured from the command line has a `base`
  `ApiSurface`.

This should not affect downstream as Android already configures it
properly and AndroidX does not track delta APIs.

Bug: 374096206
Test: ./gradlew
Change-Id: I1ba41d13c434ba5cf113a3824d0c559bab1c675e