Add optionalReporter to Codebase.Config
Moves the optional `Reporter` that can be passed through to `Codebase`
from a separate parameter into `Codebase.Config`. Although this touches
a lot of the same files as the change to wrap `AnnotationManager` in
`Codebase.Config` it is separated out because it is slightly more
complicated. That is because `Reporter` is optional in `Codebase` and
so is not passed through every path that `Codebase.Config` is.

Previously, `DefaultCodebase.reporter` would fail if a `Reporter` has
not been provided. That logic is moved into `Codebase.Config` to
maintain the previous behavior.

The `ApiFile.parseApi(...)` methods were not passed a `Reporter` so
they created a `BasicReporter`. This changes that slightly as now they
are passed the same `Reporter` as is used in the main command. That
should not cause any changes in behavior as it is not generally used on
`Codebase`s created from signature files but if it is used then it will
provide better reporting.

Some subcommands that used `ApiFile.parseApi(...)` directly relied on
that creation of `BasicReporter`. They have been modified to pass their
own instance through.

Bug: 375208805
Test: ./gradlew
Change-Id: Ibd5cca638cbeb9dd977e126ab23fc4843c698a32