Pass the whole TestFixture through to test runners
Previously, the `TestFixture` properties were copied into identical
properties in `ModelSuiteRunner.TestInputs` before passing to a test
runner. This change just passes the whole `TestFixture` through. That
will simplify the process of adding new properties to `TestFixture`.

Both runners create a `Codebase.Config` from the `TestFixture` so this
dedups that by adding a `codebaseConfig` property to `TestFixture` that
is created from the other properties.

Bug: 375182859
Test: ./gradlew
Change-Id: Ia8c197f16e658e18974f40353a85d0eb556d0349