Update selectedApiVariants when parsing signature files
Previously, a `Codebase` created by parsing signature files would not
update the `SelectableItem.selectedApiVariants` property. This change
uses the `SignatureFile.apiVariantFor(ApiSurfaces)` to mark which
`SelectableItem`s belong to which `ApiVariant`s.

It uses the `SignatureFile.forTest(List<File>)` method added in a
preceding test to create the `SignatureFile`s that are processed. That
will set the `ApiVariant` to which the `SignatureFile` belongs based on
the file path.

Bug: 374096206
Test: ./gradlew
Change-Id: I6d09f04acd734483dfb0672a1d09b934ea9c486e