Stop calling filter reference on TypeParameterItem
Previously, when checking for references to hidden classes in a
`VariableTypeItem` it would apply `filterReference` predicate to a
`TypeParameterItem`. Filter reference predicates depend on the
`ApiVariantSelectors` but that is fixed to always visible for
`TypeParameterItem`s as they cannot be separately selected from their
owner. As a result the filter reference will always assume that a
`TypeParameterItem` is included. That means that will never detect a
hidden class.

That does not cause any problems because `VariableTypeItem`s simply
reference a `TypeParameterItem` and while its type bounds can reference
hidden classes they are checked elsewhere.

Bug: 375601199
Test: ./gradlew
Change-Id: I6ba2b38a9576af64b2cb6951d65f301210850063