Tweak report filtering in ApiLint to handle SelectableItem
Previously, the `FilteringReporter.report(...)` method would filter out
issues that are reported on `Item`s which are not in the API. That
required applying a `FilterPredicate` to an `Item`. That would be
broken by the follow-up change which switches `FilterPredicate` from
`Predicate<Item>` to `Predicate<SelectableItem>`. This change refactors
that code to make it clear that the `Item` was actually a
`SelectableItem` preventing the follow-up change from breaking it.

Bug: 375699204
Test: ./gradlew
Change-Id: I8f8f0ef578910a6a5cd9bf9558680c2b25ff8418