Stop tracking added/removed parameters in ComparisonVisitor
`ComparisonVisitor` cannot actually track whether a `ParameterItem` has
been added or removed. That is because they are owner by `CallableItem`
and two `CallableItem`s are deemed to be the same if they have the
same number of `ParameterItem`s with the same types. So, if a
`ParameterItem` is added or removed then it is treated as if the
`CallableItem` with the old number of parameters has been removed and a
`CallableItem` with the new number of parameters has been added. This
change treats the adding/removal of parameters as an error showing that
the code path is never taken.

Follow-up changes will clean the code up to remove the error code path
for `ParameterItem`s.

Bug: 375739126
Test: ./gradlew
Change-Id: Ia1e4d938ebc9ef9829724cf82679249ecff590bb