Add SignatureFile.apiVariantFor(ApiSurfaces)
When loading a `Codebase` from a set of signature files it is necessary
for each `SelectableItem` to keep track of which `ApiVariant`s it
belongs. That requires that each signature file is associated with a
specific `ApiVariant` so each `SelectableItem` loaded from that file
can be associated with that `ApiVariant`.

This change adds an `ApiVariantType` to the `SignatureFile`, which is
used by `apiVariantFor(ApiSurfaces)` (in combination with
`forMainApiSurface`) to retrieve the associated `ApiVariant` for the
`SignatureFile`. A follow-up change will use that to track the
`selectedApiVariant`s.

The `ApiVariant` is retrieved at use time rather then when constructing
the `SignatureFile` because the `ApiSurfaces` may not be available at
the time the `SignatureFile` is constructed.

Bug: 374096206
Test: ./gradlew
Change-Id: I14dfaee40a488feee4135286fbbd45585fe170e3