Add stripJavaLangPrefix to TypeItem.toTypeString(...)
Previously, the stripping of `java.lang.` prefix was done after the
string representation was created which meant it could only safely
remove the prefixes from the start of the string. This change moves
the stripping into the `toTypeString(...)` implementation based on a
new `stripJavaLangPrefix` configuration property that can be set to
`NEVER` (the default), `LEGACY` (legacy behavior), and `ALWAYS`.

It continues to use the `stripJavaLangPrefix(...)` method to support
the `LEGACY` behavior as replicating that is tricky and will be done in
a follow-up change. The tests added as part of this will ensure that
is done correctly.

Other uses of `stripJavaLangPrefix(...)` are replaced with this.

The `shortenTypes(...)` method is only called from `SignatureWriter`
and is called on a string returned from `toTypeString(...)`. So, this
change passes the correct setting for `stripJavaLangPrefix` to the
`toTypeString(...)` method and removes the call to
`stripJavaLangPrefix(...)`. That broke a test for `shortenTypes(...)`
so the breaking assertions were removed as they are no longer necessary
and are covered by the new tests added by this.

The intention is that a `stripJavaLangPrefix` configuration property
will be added to `FileFormat` and used to control the behavior (like is
done for other properties) which will make it easier to migrate
existing signature files to use it.

Bug: 321061298
Test: ./gradlew
Change-Id: Ic0db33b9e173164f473b5e555749d34abbf41611