Dedup computation of array suffix to use in toTypeString(...)
When generating the string representation of an array there are two
different paths depending on whether annotations are being included or
not as placing annotations in the correct location requires some extra
work. Previously, each path would compute whether the type was a
varargs and so needed to use `...` or if it could just use `[]`. This
change dedups that computation to make a follow-up change that allows
that to be controlled through a `TypeStringConfiguration` property
simpler.

Bug: 321061298
Test: ./gradlew
Change-Id: I1f4f14a4b2af4d1134138f2352407cd75a3f1213