Remove assertions about MAGIC_VERSION used in API generator test
ag/29522183 removed the function ApiConstraint.SdkApiConstraint.isValidApiLevel. This function was previously used to verify that the MAGIC_VERSION_INT used by the API generator test was less than the MAX_LEVEL possible for an API level. However, this will no longer be necessary because the new ApiConstraint implementation doesn't use a bitset so any int can be used as an API level.

Metalava is still using a lint version with the previous ApiConstraint implementation, but since the validity check was only in a test it is safe to remove already.

This removes the isValidApiLevel check and updates the HIGHEST_KNOWN_API check to use ApiConstraints, because the new ApiVersion implementation packs both the major and minor version in one int, so I think it may be the case that a higher int doesn't necessarily make a higher API version.

Bug: 377266784
Test: `./gradlew test`. The lint version with the ApiConstraint updates hasn't been released yet, so that will be tested by the integration branch, but this now only uses functions that exist in both versions.
Change-Id: I78d018bb6183bd31bdd44da075bd8441c309d834