Create a single Codebase for flagged API reversion
The previously released API against which compatibility checks are
performed is split into two separate `Codebase`s, one for the released
API and one for the removed API. Previously, they were iterated over by
the flagged API functionality to find the previously released item when
reverting a change but it did not differentiate between the released
and removed APIs.

This change creates a single `Codebase` containing the released API and
removed API. That will make follow-up work that tracks `ApiVariant`s
simpler as it only has to deal with a single `Codebase` and not a list.
Longer term the compatibility checks will also be done on a single
`Codebase` and use `selectedApiVariants` to differentiate between
removed and released API where necessary.

Bug: 374096206
Test: ./gradlew
Change-Id: I6dd689eb5ef432d212a7ac96c7f42649c0a5194c