Update baseline key generation for kt source elements
This updates psi baseline key generation to work with `KtFunction`s and to handle top-level `KtProperty`s.

These cases do not frequently (or possibly ever?) come up in issue reporting, but the following change adds more detail to an error in psi type creation, and they do occur in that context.

Bug: 377315363
Test: Added PsiFileLocationTest to verify baseline keys for these cases
Change-Id: I5495e23448469cb158dbbbe4a523ddaaf6343933